api/pyproject.toml
[tool.poetry]
name = "minitwitter"
version = "0.0.1"
description = ""
authors = ["Breno Viana <bmfviana@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
# Python
python = ">=3.9,<3.10"
# Environment
python-decouple = "3.4"
# Django + Django REST
django = "4.2"
djangorestframework = "3.14"
djangorestframework-simplejwt = "5.3.1"
drf-spectacular="0.27.2"
# Database
psycopg2-binary = "2.9.*"
pillow="11.0.0"
django-redis = "5.4.0"
# Celery
celery={ version = "^5.4.0", extras = ["redis"] }

[tool.poetry.group.dev.dependencies]
factory_boy="3.3.1"
freezegun="1.2.2"

[tool.poetry.group.ci-fmt.dependencies]
ruff = "0.7.0"
black = "24.10.0"

#

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | venv
    | env
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.ruff]

[tool.ruff.lint]
extend-select = [
    "F", # Pyflakes
    #	"E", # Pycodestyle
    "W",  # Pycodestyle
    "I",  # isort
    "UP", # pyupgrade
    #	"D",   # pydocstyle
    #	"B", # flake8-bugbear
    #	"SIM", # flake8-simplify
]

[tool.ruff.lint.pydocstyle]
convention = "google"
